requirements.txt
# Data Ingestion & Wrangling
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
pyarrow>=14.0.0

# Feature Engineering
scikit-learn>=1.4.0
feature-engine>=1.6.0

# ML Models
xgboost>=2.0.0
lightgbm>=4.0.0

# Optimization
scipy>=1.11.0

# Explainability
shap>=0.44.0

# Experiment Tracking
mlflow>=2.10.0

# Monitoring
evidently>=0.4.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.0
httpx>=0.26.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Linting
ruff>=0.2.0

# Config
pyyaml>=6.0.0

# Utilities
joblib>=1.3.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
